requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
PyGithub>=2.0.0
pydantic>=2.0.0
gitpython>=3.1.30
python-dotenv>=1.0.0
requests>=2.28.2
cryptography>=40.0.0
